Compatible with different system versions of ssl.
